refact: factor out participant criteria component
